Potential fix for pull request finding 'CodeQL / Implicit narrowing conversion in compound assignment'

+++ src/main/java/io/github/joachimvn/ui/tournament/TournamentSummary.java
@@ -147,7 +147,7 @@ final class TournamentSummary {
         double fastestAvg = Double.MAX_VALUE;
         double slowestAvg = 0;
         double totalMoveSum = 0;
-        int totalMoveCount = 0;
+        long totalMoveCount = 0;
         for (Difficulty d : tableItems) {
             long[] mt = moveTotals.get(d);
             if (mt == null || mt[1] == 0) continue;
